[lib][trusty] smp and 64 bit fixes

+++ lib/trusty/include/lib/trusty/ipc_msg.h
@@ -56,15 +56,15 @@ typedef struct ipc_msg_kern {
 } ipc_msg_kern_t;
 
 typedef struct ipc_msg_user {
-	uint		num_iov;
+	uint32_t	num_iov;
 	user_addr_t	iov;
 
-	uint		num_handles;
+	uint32_t	num_handles;
 	user_addr_t	handles;
 } ipc_msg_user_t;
 
 typedef struct ipc_msg_info {
-	size_t		len;
+	uint32_t	len;
 	uint32_t	id;
 } ipc_msg_info_t;
